+++ autoreject/utils.py
@@ -24,6 +24,19 @@ def _get_ch_type_from_picks(picks, info):
     return keys
 
 
+def _check_ch_locs(chs):
+    """Check if channel locations exist.
+    Parameters
+    ----------
+    chs : dict
+        The channels from info['chs']
+    """
+    locs3d = np.array([ch['loc'][:3] for ch in chs])
+    return not ((locs3d == 0).all() or
+                (~np.isfinite(locs3d)).all() or
+                np.allclose(locs3d, 0.))
+
+
 def _check_data(epochs, picks, ch_constraint='data_channels',
                 verbose='progressbar'):
     BaseEpochs = _get_epochs_type()
@@ -33,6 +46,10 @@ def _check_data(epochs, picks, ch_constraint='data_channels',
     if epochs.preload is False:
         raise ValueError('Data must be preloaded.')
     n_bads = len(epochs.info['bads'])
+
+    if not _check_ch_locs(epochs.info['chs']):
+        raise RuntimeError('Valid channel positions are needed'
+                           'for autoreject to work')
 
     picked_info = mne.io.pick.pick_info(epochs.info, picks)
     ch_types_picked = {