+++ doc/whats_new.rst
@@ -24,6 +24,8 @@ Bug
 
 - Fixed bug in picking bad channels during interpolation. This bug only affects users who got an assertion
   error when running :class:`autoreject.Autoreject`. Fixed by `Mainak Jas`_ in `#115 <https://github.com/autoreject/autoreject/pull/115>`_
+- Added check for channel locations so that autoreject does not
+  hang when the channel positions are nan. Fixed by `Mainak Jas`_ in `#130 <https://github.com/autoreject/autoreject/pull/130>`_
 
 API
 ~~~